notifications-service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
aio-pika==9.4.3
pydantic==2.10.3
PyJWT==2.10.1
python-dotenv==1.0.1
